+++ lib/d4rt.dart
@@ -43,3 +43,4 @@ export 'src/utils/extensions/iterable.dart';
 export 'src/runtime_interfaces.dart';
 export 'package:d4rt/src/async_state.dart';
 export 'package:d4rt/src/utils/logger/logger.dart';
+export 'package:d4rt/src/security/permissions.dart';